requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
python-docx>=1.1.0
kaleido>=0.2.1
